pyproject.toml
[project]
name = "dl-image"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.15.0",
    "timm>=0.9.0",
    "matplotlib>=3.7.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "pillow>=10.0.0",
    "scikit-learn>=1.7.2",
]
